refactor: (test cli) SuiteMetric members total_runs and total_failures switched from int type to size_t type

+++ src/ctest_cli/executor.h
@@ -6,8 +6,8 @@
 #define CTEST_MAX_LINE_LEN 512
 
 typedef struct {
-  int total_runs;
-  int total_failures;
+  size_t total_runs;
+  size_t total_failures;
   int crashed;
 } SuiteMetrics;
